components/runners/claude-code-runner/pyproject.toml
[project]
name = "claude-code-runner"
version = "0.2.0"
description = "AG-UI Server for Claude Code SDK"
readme = "CLAUDE.md"
requires-python = ">=3.11"
authors = [
  { name = "vTeam" }
]
dependencies = [
  # AG-UI Server
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.23.0",
  
  # AG-UI Protocol SDK (official)
  "ag-ui-protocol>=0.1.0",
  
  # Claude SDK
  "anthropic[vertex]>=0.68.0",
  "claude-agent-sdk>=0.1.23",
  
  # Observability
  "langfuse>=3.0.0",
  
  # Utilities
  "pydantic>=2.0.0",
  "aiohttp>=3.8.0",
  "requests>=2.31.0",
  "pyjwt>=2.8.0",
  
  # MCP integrations
  "mcp-atlassian>=0.11.9",
]

[tool.uv]
dev-dependencies = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.1.0",
  "black>=23.0.0",
  "httpx>=0.24.0",
]

[tool.setuptools]
py-modules = ["main", "adapter", "auth", "config", "context", "mcp", "observability", "prompts", "security_utils", "utils", "workspace"]
packages = ["tools", "ag_ui_claude_sdk", "endpoints", "middleware", "ambient_runner", "ambient_runner.endpoints", "ambient_runner.middleware", "ambient_runner.platform", "ambient_runner.bridges"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
